# web/packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teggo B2B Commerce API
  version: 0.1.0
  description: >
    The single source of truth for the Teggo API contract. The TypeScript
    client used by both frontends is generated from this file (do not hand-edit
    the generated schema). Grown module-by-module as the backend lands.
servers:
  - url: http://localhost:8080
    description: Local Go API
...
components:
  securitySchemes:
    adminAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    storefrontAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    vendorAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Shipment:
      type: object
      required: [id, public_id, order_id, status]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        order_id: { type: integer, format: int64 }
        carrier: { type: string, nullable: true }
        tracking_number: { type: string, nullable: true }
        status: { type: string, enum: [pending, shipped, delivered, returned] }
        shipped_at: { type: string, format: date-time, nullable: true }
        warehouse_id: { type: integer, format: int64, nullable: true }
...
    ShipmentInput:
      type: object
      required: [items]
      properties:
        carrier: { type: string, nullable: true }
        tracking_number: { type: string, nullable: true }
        warehouse_id: { type: integer, format: int64, nullable: true }
        items:
          type: array
          items:
            type: object
            required: [order_item_id, quantity]
            properties:
              order_item_id: { type: integer, format: int64 }
              quantity: { type: string }
    ShipmentStatusPatch:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, shipped, delivered, returned] }
    InvoiceItem:
      type: object
      required: [id, description, quantity, unit_price, row_total]
      properties:
        id: { type: integer, format: int64 }
        description: { type: string }
        quantity: { type: string }
        unit_price: { type: string }
        tax_amount: { type: string }
        row_total: { type: string }
    ReturnItem:
      type: object
      required: [id, order_item_id, product_id, quantity, sku, name, unit_price]
      properties:
        id: { type: integer, format: int64 }
        order_item_id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        quantity: { type: string }
        reason: { type: string, nullable: true }
        sku: { type: string }
        name: { type: string }
        unit_price: { type: string }
    CreditNote:
      type: object
      required: [id, public_id, amount, currency, status]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        invoice_id: { type: integer, format: int64, nullable: true }
        amount: { type: string }
        currency: { type: string }
        status: { type: string, enum: [draft, issued, void] }
    ReturnSummary:
      type: object
      required: [id, public_id, order_id, customer_id, status]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        order_id: { type: integer, format: int64 }
        customer_id: { type: integer, format: int64 }
        status: { type: string, enum: [requested, approved, rejected, received, closed] }
        reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    ReturnDetail:
      type: object
      required: [id, public_id, order_id, customer_id, status, items, credit_notes]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        order_id: { type: integer, format: int64 }
        customer_id: { type: integer, format: int64 }
        status: { type: string, enum: [requested, approved, rejected, received, closed] }
        reason: { type: string, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/ReturnItem' } }
        credit_notes: { type: array, items: { $ref: '#/components/schemas/CreditNote' } }
    ReturnInput:
      type: object
      required: [items]
      properties:
        reason: { type: string, nullable: true }
        items:
          type: array
          items:
            type: object
            required: [order_item_id, quantity]
            properties:
              order_item_id: { type: integer, format: int64 }
              quantity: { type: string }
              reason: { type: string, nullable: true }
    ListWrapperReturn:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ReturnSummary' } }
    ListWrapperCreditNote:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CreditNote' } }
    AgingInvoice:
      type: object
      required: [public_id, customer_id, status, grand_total, currency, days_overdue, bucket]
      properties:
        public_id: { type: string, format: uuid }
        customer_id: { type: integer, format: int64 }
        status: { type: string }
        grand_total: { type: string }
        currency: { type: string }
        due_at: { type: string, format: date-time }
        days_overdue: { type: integer }
        bucket: { type: string, enum: [current, '1-30', '31-60', '61-90', '90+'] }
    ARAgingReport:
      type: object
      required: [buckets, open_total, items]
      properties:
        buckets: { type: object, additionalProperties: { type: string } }
        open_total: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/AgingInvoice' } }
    OverdueSweepResult:
      type: object
      required: [marked_overdue]
      properties:
        marked_overdue: { type: integer }
    InvoiceDetail:
      type: object
      required: [id, public_id, status, currency, subtotal, tax_total, grand_total, items]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        customer_id: { type: integer, format: int64 }
        status: { type: string, enum: [draft, issued, paid, overdue, void] }
        currency: { type: string }
        subtotal: { type: string }
        tax_total: { type: string }
        grand_total: { type: string }
        due_at: { type: string, format: date-time, nullable: true }
        pdf_url: { type: string, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/InvoiceItem' } }
    InvoiceSummary:
      type: object
      required: [id, public_id, order_id, status, currency, grand_total]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        order_id: { type: integer, format: int64 }
        status: { type: string }
        currency: { type: string }
        subtotal: { type: string }
        tax_total: { type: string }
        grand_total: { type: string }
        issued_at: { type: string, format: date-time, nullable: true }
        due_at: { type: string, format: date-time, nullable: true }
        pdf_url: { type: string, nullable: true }
    Payment:
      type: object
      required: [id, public_id, customer_id, method, amount, currency, status]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        invoice_id: { type: integer, format: int64, nullable: true }
        order_id: { type: integer, format: int64, nullable: true }
        customer_id: { type: integer, format: int64 }
        method: { type: string, enum: [card, ach, invoice, po, mpesa] }
        gateway: { type: string, nullable: true }
        gateway_reference: { type: string, nullable: true }
        amount: { type: string }
        currency: { type: string }
        status: { type: string, enum: [pending, authorized, captured, failed, refunded] }
        captured_at: { type: string, format: date-time, nullable: true }
    PaymentInput:
      type: object
      required: [customer_id, method, amount, currency]
      properties:
        invoice_id: { type: integer, format: int64, nullable: true }
        order_id: { type: integer, format: int64, nullable: true }
        customer_id: { type: integer, format: int64 }
        method: { type: string, enum: [card, ach, invoice, po, mpesa] }
        gateway: { type: string, nullable: true }
        gateway_reference: { type: string, nullable: true }
        amount: { type: string }
        currency: { type: string }
        status: { type: string, enum: [pending, authorized, captured, failed, refunded] }
    ListWrapperShipment:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Shipment' } }
    ListWrapperInvoiceSummary:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/InvoiceSummary' } }
    ListWrapperPayment:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Payment' } }
...
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /admin/orders/{id}/shipments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [OrderToCash]
      operationId: adminListShipments
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperShipment' } } } }
    post:
      tags: [OrderToCash]
      operationId: adminCreateShipment
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ShipmentInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Shipment' } } } }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /admin/shipments/{id}/status:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    patch:
      tags: [OrderToCash]
      operationId: adminPatchShipmentStatus
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ShipmentStatusPatch' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Shipment' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /admin/orders/{id}/invoices:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [OrderToCash]
      operationId: adminListInvoicesForOrder
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperInvoiceSummary' } } } }
    post:
      tags: [OrderToCash]
      operationId: adminIssueInvoice
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceDetail' } } } }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /admin/invoices/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [OrderToCash]
      operationId: adminGetInvoice
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /admin/invoices/{id}/pdf:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [OrderToCash]
      operationId: adminRegenerateInvoicePdf
      security: [{ adminAuth: [] }]
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { type: object, properties: { enqueued: { type: boolean } } } } } }
  /admin/invoices/{id}/payments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [OrderToCash]
      operationId: adminListInvoicePayments
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperPayment' } } } }
  /admin/payments:
    post:
      tags: [OrderToCash]
      operationId: adminRecordPayment
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } } }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /admin/payments/{id}/refund:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [OrderToCash]
      operationId: adminRefundPayment
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/invoices:
    get:
      tags: [OrderToCash]
      operationId: storefrontListInvoices
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperInvoiceSummary' } } } }
  /storefront/invoices/{publicID}:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [OrderToCash]
      operationId: storefrontGetInvoice
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/invoices/{publicID}/pay:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [OrderToCash]
      operationId: storefrontPayInvoice
      summary: Pay an invoice by card via the configured gateway
      security: [{ storefrontAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string, description: 'Gateway card token / nonce (no raw card data)' }
      responses:
        '200': { description: Paid, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceDetail' } } } }
        '402': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/orders/{publicID}/pay:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [OrderToCash]
      operationId: storefrontPayOrder
      summary: Pay an order by card (issues its invoice first if none exists yet)
      security: [{ storefrontAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string, description: 'Gateway card token / nonce (no raw card data)' }
      responses:
        '200': { description: Paid, content: { application/json: { schema: { $ref: '#/components/schemas/InvoiceDetail' } } } }
        '402': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /admin/invoices:
    get:
      tags: [OrderToCash]
      operationId: adminListInvoices
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperInvoiceSummary' } } } }
  /admin/invoices/aging:
    get:
      tags: [OrderToCash]
      operationId: adminInvoiceAging
      summary: AR-aging report — open invoices bucketed by days past due.
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ARAgingReport' } } } }
  /admin/invoices/overdue-sweep:
    post:
      tags: [OrderToCash]
      operationId: adminRunOverdueSweep
      summary: Flip past-due issued invoices to overdue and dun the customers.
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OverdueSweepResult' } } } }
  /admin/returns:
    get:
      tags: [OrderToCash]
      operationId: adminListReturns
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperReturn' } } } }
  /admin/returns/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [OrderToCash]
      operationId: adminGetReturn
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /admin/returns/{id}/approve:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
    post:
      tags: [OrderToCash]
      operationId: adminApproveReturn
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /admin/returns/{id}/reject:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
    post:
      tags: [OrderToCash]
      operationId: adminRejectReturn
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /admin/returns/{id}/receive:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
    post:
      tags: [OrderToCash]
      operationId: adminReceiveReturn
      summary: Mark received — restock items and issue a credit note.
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /admin/orders/{id}/returns:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
    get:
      tags: [OrderToCash]
      operationId: adminListOrderReturns
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperReturn' } } } }
    post:
      tags: [OrderToCash]
      operationId: adminCreateReturn
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDetail' } } } }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /admin/credit-notes:
    get:
      tags: [OrderToCash]
      operationId: adminListCreditNotes
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperCreditNote' } } } }
  /storefront/returns:
    get:
      tags: [OrderToCash]
      operationId: storefrontListReturns
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperReturn' } } } }
  /storefront/orders/{publicID}/returns:
    parameters: [{ name: publicID, in: path, required: true, schema: { type: string, format: uuid } }]
    post:
      tags: [OrderToCash]
      operationId: storefrontCreateReturn
      summary: Buyer self-serve return request for their own order.
      security: [{ storefrontAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/ReturnDetail' } } } }
        '422': { $ref: '#/components/responses/ErrorResponse' }
...